Update WLT_Project.Target.cs
In this expanded version, I have added comments to explain what each section of the code does. The comment at the top reminds you to fill out the copyright notice in the Project Settings page.

I have also added a comment above the constructor to indicate that it is the constructor for the WLT_ProjectTarget class. Finally, I have added a comment to explain what the ExtraModuleNames property does. These comments should make it easier for you or anyone else to understand the purpose of the code.