Update WLT_ProjectGameModeBase.h
n this version of the code, I have added comments to explain what each section of the code does. The comment at the top reminds you to fill out the copyright notice in the Project Settings page.

I have also added a comment block above the class declaration to provide more information about what the class does. In this case, I explained that it is the base class for the game mode in the WLT_Project game and that it provides additional functionality specific to the game. This information should make it easier for you or anyone else to understand the purpose of the code.