Update WLT_ProjectGameModeBase.cpp
In this version of the code, I have added comments to explain what each section of the code does. The comment at the top reminds you to fill out the copyright notice in the Project Settings page.

I have also added a comment block above the implementation file to explain that it contains the definitions for the functions and member variables declared in the header file. I also provided a brief explanation of what the implementation file is for.

Finally, I left a comment for you to define any member functions for the AWLT_ProjectGameModeBase class. This is where you would add the actual code for any functions that you declared in the header file.