Update WLT_Project.h
In this version of the code, I have added comments to explain what each section of the code does. The comment at the top reminds you to fill out the copyright notice in the Project Settings page.

I have also added a comment block above the class declaration to explain that it is the header file for the WLT_Project class. In this block, I explained that it includes any necessary headers and declares any classes, functions, and variables used in the class.

Next, I added member function declarations for StartGame and EndGame functions, as well as a default constructor and destructor. These are just examples, you can replace them with any member functions that you need for your class.

Finally, I declared a private member variable Score, which is an example of a member variable that you might use in your WLT_Project class. Again, you can replace this with any member variables that you need for your class.